pyproject.toml
[project]
name = "open-chemistry-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.2",
    "pyscf>=2.10.0",
    "rdkit-pypi>=2022.9.5",
    "requests>=2.32.4",
    "simpletooling @ git+https://github.com/EvoEvolver/simpletooling.git",
]
